Update Help text for research_summary on Credentialing Application
Currently, during credentialing, we have a minimum limit of 20 words(but this is not obvious to the user as this information is not displayed to user). Users only see the message to provide the details about their research. So when trying to submit credentialing application with less than 20 words, they will keep getting served error by validators(but the error doesnot make it clear about the issue)

As pointed out by January, it will be a good idea if we could show the minimum word limit explicitly.