refactor template and views to use the SafeHtmlField
The reason that i decided to make a separate page for event edit(compared to previous model) is because i was not able to get ckeditor to work with modal properly. I found out that when there was two modal(add event and edit event modal) ckeditor was working for the first modal only.

I had to refactor the view, because previously the view `update_event` was only accepting post request, but now since we have a separate page for event, we need it to work with GET requests also