update model to save effort in future
It is obvious that in future, we will add more access mechanism to event dataset. So as suggested by @alistairewj, it might be a good idea to make the length of access mechanism fullproof by increasing it to a safe number.

With max_length=10, we should be able to add new mechanism without needing to change the model in future.