use unique names for navigation buttons
There are a few navigation links which use the same id,
presumably due to copy paste. This changes them to use
a consistent naming scheme: nav_*. This change does not
seem to have a functional impact as these IDs are not
used anywhere AFAIK.

Also changed the code_of_conduct_nav ID to nav_code_of_conduct,
for similar consistency reasons and to avoid overloading the
boolean variable.