remove text assertion from test_views
as suggested here https://github.com/MIT-LCP/physionet-build/pull/1876#discussion_r1113709658
with assertion of text in the template, we will need to maintain this in future, if the error messages are updated for EventDataset.
So to save hassale in future, we decided to avoid using the text assertion.

Currently we are validating that the form returned error or the event agreement creation/update was not successful by checking status code and template used.