add `has_access` method to EventDataset
This method is independent of the PublishedProject.has_access method. it will only check if the user should
have access to the dataset for the event.(so it can be used in the event detail page to check if the user has access)

It is expected that the PublishedProject.has_access method will use this method to check if the user has access
in case the datasetaccess is made through an event.