fix Event description field not showing line breaks
It looks like on one of the previous PRs i had applied a class "row" by mistake to the event description field.
This was causing the line breaks to not show up(because of the class "row" adding `display:flex;` css property, which messed up the displayed html format ).

I removed the class and now the line breaks are showing up properly

Note: i was able to reproduce the issue on the cases where the paragraphs were very short, like 1 or 2 words(i missed this case when testing the previous PR because of that)