event_entries.html: make checkboxes disabled.
The input elements that indicate "cohost" status don't do anything,
since the backend view along with the accompanying script has been
removed.

For the time being (until this functionality can be re-implemented in
a way that is secure and follows the HTTP standard), make these
checkboxes disabled, so as not to give visitors the impression that
they do anything.