Fix Validation error on EventAgreement Edit
On https://github.com/MIT-LCP/physionet-build/pull/1876, we introduced a bug on Event Agreement Edit. The bug is if you try to edit an Event Agreement, it will throw a validation error saying the "Name" and "Version" already exists.
The error comes from `EventAgreementForm`'s `clean` method. Here the if statement was throwing validation error if the name, version already exists as a row in database.
The logic is wrong as it doesnot account for that fact that when you update an Event Agreement, the event agreement's name and version might be the same as before.

So in case of update/edit, we need to check if the existing row in the database is from the event agreement being edited; if yes, we shouldnot throw an error.