Update Event Agreement Test
One of the reason we had the bug, was because on the event agreement edit test, we had a test which was editing the name of the event agreement. Hence the test didnot catch the error.

I added a new test that updates the `html_content` of the event agreement.