update EventAgreement Model __str__ method
The Option displayed on the `EventForm` for the field `event_agreement` of model `Event` depends on `__str__` method of `EventAgreement`. Since we can have different version of same `EventAgreement`, it is a good idea to show the `EventAgreement` version to the Event Host.

Before the change, if there are 2 Event Agreement with Name "My Event Agreement" but different versions, then the Event Host will see the same name repeated twice which will be confusing.

PS: i also checked how DUAs are implemented in project, and i think the expectation there might be that the Administrator will set the old DUAs to inactive.

So this commit might be optional if we assume
1. Old version of event Agreement will not be used by Event Host.
2. Platform Administrator will set the old versions of Event Agreement to inactive.