update `has_access` method to check for event_agreement
Note: Although we have plans to refactor this method in the future, i had to update it for now because we dont know how we will refactor it later. Since this change is required for participant to access the dataset through events, it should be okay to update it here. The logic will get refactored when we refactor the function.