refactor `get_application_to_be_rejected` to remove redundancy
The function `get_application_to_be_rejected` gives us a list of credentialing applications which have pending reference response that needs to be rejected.

and for this PR, we need a nearly identical function, with just difference in the name of function and name of variables. For example for sending a reminder, we will need a function `get_application_to_be_reminder` and it will end up having the exact code.

So to prevent that i created a generic function that takes days, and number and returns the filtered applications. This function should work for the existing code and will also work for sending reminders.