CredentialApplication: add field reminder_datetime
Because we don't want to keep sending reminder to the same person every hour, We need to be able to track if we had sent a reminder to a user for the credentialapplication. For this we need to store something in the database.

I couldn't find any kind of model to store Log or notifications on physionet-build and creating one will be  out of scope of this PR. So i decided to add a field here.