Use DjangoEnum for status in CredentialApplication
On PR https://github.com/MIT-LCP/physionet-build/pull/1913, we discussed the current system is not human readable because of the numbers.

This commit will add a constraint check in database for the status field, but everything else should be the same. All the existing code should function as it is.