CredentialApplicationStatus -> Status
Refactoring the class name for CredentialApplication status.
the `CredentialApplicationStatus` is kind of big and doesnot add any value for its length(as this class is already under the `CredentialApplication`. So changing to `Status` makes it shorter and doesnot affect the readability of code

Also since this commit is just refactoring the class name and doesnot change anything in database, there is no change in migration file

Thanks to @bemoody