Display permission groups on usermanagement page
As discussed on https://github.com/MIT-LCP/physionet-build/issues/1548, We want to show the permission/group permissions that a user has on the admin console.

This commit adds the names of the groups that the user belongs to. In followup commits i will add a view that will list all permissions that a group has and also list the people who are in that permission group.