Handle bucket root correctly when copying GCS objects
GCSObject has an edge case, where the bucket root's path is turned into a
forward slash to make it behave like all other "folders". This has
historically been an issue since GCS paths do not contain leading
slashes and are always relative to the bucket name. Logically, the slash
is a *trailing* slash, i.e. the object name is an empty string (root of
bucket) and to make it seem like a folder, the slash is appended,
resulting in "/". In the future this might be refactored by not adding
the slash for root and reimplementing `is_dir`. This might have other
consequences so it should be taken with consideration.