migration file:create token in existing application
Here, for credential applications where the reference has not responded
we are setting the verification token.
This is because, if the admin hasnot yet sent the reference request,
we should be using the token for the new reference request generated for existing credential applications.
similarly, we just need this for the applications that have are
still in the pending state