create onplatform training with json
Added a serializer to allow admins to create a new training or update the existing training.

To track the updates to training, we are using semantic versioning. If the updated training has a major update eg from 1.9 to 2.0, then all the users who did completed the previous versions will be sent an email asking them to complete the new version of training with x days.