implement trainings for user
Allows users to take the test from settings.

The quiz.html file loads all the content and questions based on the order. All the data is hidden from user and gradually displayed as users go through the training

We have two level of verifications of answers. the first one is done on quiz.html where users go through the content and they will see the related questions and answer the questions. users can go to next section or skip the question only if they have answered the question correctly.
We also have another verification done on view. here if the users had correctly answered the questions a json string is generated and sent via POST, we will then check the database to verify that the data in json string(question,answer) matched with the database. This should help us stop someone from just sending a post request to complete the training.

Quick note:
1. Users are not able to resume the test if they refresh the browser
2. Users can take the same test multiple times
3. Only multiple choice questions is supported yet.