install-pn-test-server: make /physionet/python-env/physionet a link.
In order to upgrade Python libraries atomically, we want to be able to
change /physionet/python-env/physionet to point to a new location,
while leaving the old tree in place.

When setting up the test environment, create 'physionet' as a link
pointing to the real directory 'initial', so that subsequent scripts
can overwrite the link atomically.