test-upgrade.sh: install python libraries without upgrading.
We want to avoid "upgrading" python libraries using pip, because this
may not give predictable results, and if installation fails it can
leave the system in a broken half-installed state.  Instead, we want
to install packages from scratch whenever it is required.

To be consistent with the new behavior of the git update hook, do the
same thing when running upgrade tests: rather than "upgrading" the
virtualenv directory, delete the old one and create a new one.  This
makes caching simpler, too, since the directory contents should be
reproducible just from requirements.txt.