Add Training Module
Currently the training lacks proper organization. We have a OnPlatform Training which holds bunch of content and quizzes under it directly.

On all online learning platforms or physical books, the training/courses are organized under chapter(in books) and modules/lessons and sections. With the idea of having modules under training instructors will be able to organize the training content under appropriate modules.

Also, i didnot squash this commit just in case we want to revert the changes later.