CredentialApplication.get_review_status:refactor
Currently, the  status returned by CredentialApplication.get_review_status is 'Awaiting a response from your reference' which can be confusing as this method is not used for Event to show EventHost the detail of credentialing application.

Removing "Your" should not make a difference to the existing users/admin as `Awaiting a response from reference`  is still relevant.