Implement modules on training
Here i created basic links that lets users access the modules, take the training, and submit the modules after completion.

The idea/expectation is that users need to complete the modules sequentially(enforced in code). So when the user submits the last module, we will provide them a certificate.

To complete this, we need to track the progress of the users(which is something we already want). So in the coming few commits, i will add the resume feature and complete TODO's of this commit(Since this commit depends on few unwritten features, i left notes)