Implement Resume Feature step 2
This commit implements 2nd part of resume feature. Here we will mark the progress of users as they go through the content and quizzes, so that next time they reload the module, they will be redirected to the last section they left

When a user clicks next button, then we will temporarily disable the next button and send a post request to backend to update the progress. if the server responds with success we will display next slide. If the server responds with error, we will show an error message and ask them to try again. And finally enable the next button.