Create Model for Course,tracking user progress
1. Course

The idea is that a Training Course will be defined with user.TrainingType. This is the ultimate(top) model for a training course.
The Course content for TrainingType model is implemented by training app. On Training app, `Course` model
can be created for each TrainingType. For different versions of the same training course, we can create as many `Course`
models as we want as long as the version is different.

A Course is divided into modules. Each module has a description and a list of contents and quizzes. Modules are like
chapters in a book. Each module has a list of contents and quizzes. Contents are like paragraphs in a chapter. Quizzes are
like questions in a chapter.
for ordering the modules, contents and quizzes, we have used `order` field. This field is used to order the modules, contents and quizzes
The ordering is unique for each instance of the parent model. It is expected that the order will start from 1 and will be incremented by 1
with no gaps.

2. Tracking User Progress during training

When a user starts a training, a `CourseProgress` model should be created for that user and the version of course.
This model tracks the progress of the user during the course. Similarly, when a user starts a module, a ModuleProgress
model should be created for each module in the training. For quiz, content progress, we should create a instance of
 CompletedContent or CompletedQuiz model when the user completes a content or quiz.
I dont think we need to track when someone started a content or quiz as they are expected to complete in few minutes.