single method for authorization
Here i created a single method for handling all authorization related stuff.
Here the idea is that we simply authorization down to this method and all
future changes can be done in this method.
The current usage of authorization  is with `project.has_access` and `project.accessible_by`(HDN)  which we can also achieve with using this new method(if we want). for example

>>> from project.authorization.access import get_accessible_projects
>>> from user.models import User
>>> user = User(username='amitupreti')
>>> queryset = get_accessible_projects(user)
>>> has_access = queryset.contains(project) #boolean result

I will also add another method that copies the `project.has_access`
in case that has a better performance.