update logic  to avoid integrity error
When testing, i found that during delete and move the pages up/down, i was getting Integrity error due to unique constraint on nav_order. Although, i cant reproduce the error consistently.

The logic for up/down should work well as the logic looks good. I made a small change to increase the temporary nav_order by 1000(just in case there is a unexpected gap between the nav_orders due to improper deletion of Static page). Example of improper deletion would be if a Static page is deleted but the nav_order was not adjusted properly, which can lead nav_orders like 1,3,4 and the move_up and move_down logic might fail in cases like that.

Also, updated the delete logic of static page to make sure that the nav_order is adjusted in a serial order.