refactor logic for Research Env
Context: The goal here is to refactor to let access controlled through DataAccess without changing the existing logic.
And the existing logic is if the project has research environment enabled, the user should not be able to view/download files(they can only create research environment and play with the data there)

In this commit, we divided the access into 2 section(first for RE and the rest). The outcome should be same as before which is if RE is enabled show the `published_project_denied_downloads.html` else show normal file access stuff