Direct Download refactor: step 2
replaces allow_file_downloads with can_view_files(this controls if a user has download access to the project by using the DataSource).

Here the idea is that all the direct download stuff will be inside
```
{% if can_view_files %}
```

and the other access methods like gcp, aws will be outside this block.

So the logic will be
1. If the project is RE project, show the RE warning message
2. Else
2.1 If the project has other access outside RE and direct download show that.
2.2 If the project has direct download access, then show direct download stuff

To make it easier to understand the changes, i am breaking down the commits to very smaller chunks

Note that 2.1 and 2.2 dont have to be either or(as currently a project can be directly accessible and also be accessible via gcs,aws etc)