gracefully handle event participation request
Fixes https://github.com/MIT-LCP/physionet-build/issues/1945

By design EventParticipant can only have 1 entry per (event,user).
The reason for the bug was that we were not checking if a user has been already approved for a event before accepting their participation request. The type of scenario can happen if Event host has two tabs open, and approves a user "a" on first tab, goes to 2nd tab and tries to approve the user "a" again(here since the Event host hasnot refreshed the page, it is not aware that the user has already been approved)

The simplest way to handle this looks like checking if the user is an active participant before attempting to approve a participation request and show an error message to Event Host.