give title highest weight for relevance
Currently, when someone searches for a project
on the platform, we provide the highest importance to
description, it doesn't look right. For example
 if I was searching on google,
 I would expect the first result to be
the ones which are spot on match like the
page content title has things i want to see.

Similarly, if a user searches for mimic, we
should be first showing project that have mimic
on the title, instead of the ones with mimic on
description