Improve relevance scoring logic
As our goal is to find the most relevant project according to search, and since we are checking for matches in 3 fields and weighting the importance of match found on each field, i think it makes sense that we make sure we use all 3 fields to calculate the relevance score. As before this, if a match was found on the title, we didn't bother to check description.