Demo Commit for customizable css theming
Healthdatanexus theme is a different from physionet
hdn has a bluish color and different gradient color and also other smaller color changes.
Currently, this is achieved by modifying the forked physionet, which
requires us to manually edit the css and copy the css files.

This commit is just a demo to get feedback on what everyone thinks about
 making the theme customizable.
If this looks acceptable, i can go ahead edit this commit message and
make the rest of relevant or all css customizable like this(so for all css,
 we would be using scss and compiling them to css.