Make required text more visible for state/province
During onboarding users to healthdatanexus platform,
January noticed that many users were confused with the
State/Province field as with a quick glance it looks
like an optional field(but is actually required for US/Canada residents).
This caused them to have to resubmit the form.

It seems like a good idea to have the Required message on the input label,
as the eye seems to see that first.