update validator for research summary
Before the change, the research summary is very restrictive,
users can only use from a list of authorized special characters.

After the change, we allow everything as long as the first character is a letter
or a number.

Summary:
r'[a-zA-Z][\w\'\[\]\n\r~!@#$%&*()+={};:‘’“”"<>?,./` -]+' (OLD)
r'\w.*' (NEW)

As discussed in https://github.com/MIT-LCP/physionet-build/pull/1980,
we want to make validators permissive unless a character poses a security risk.

Since the research_summary is not a html field, so
we wont ever render this as html by using something like
|safe in the template. Also, since we were already
allowing characters like </>, so that also kind of supports
that this seems safe.