update validator for reference response
Currently, the reference response is very restrictive,
users can only use from a list of authorized special characters.

After the change, we allow everything as long as the first character is a letter
or a number.

As discussed in https://github.com/MIT-LCP/physionet-build/pull/1980,
we want to make validators permissive unless a character poses a security risk.

Since the reference_response_text is not a html field, so
we wont ever render this as html by using something like
|safe in the template.