setup static-override folders for custom deployment
As discussed on https://github.com/MIT-LCP/physionet-build/issues/1996,
we are making the theme customizable for deployment
by .env configuration. This lead us modifying existing css files
adding configurable image files.
This folder should be used to store deployment specific
static files.

It is expected that compilestatic will be ran first(which will
generate the css, image files and store them to the static-override folder),
and then collectstatic will be ran which will collect static files
from static and static-override directories

Note that, the `static-overrides` comes before the `static` on base.py
This is intentional to make sure `static-overrides` is given
first priority if there are same file in these two directories.