add model validator
Data Location and AccessMechanism are tightly coupled.

For example, as of now Research Environment is only implemented for
GOOGLE_CLOUD_STORAGE, or Direct data access is only available for
the resources which are directly stored on server.

The validator will first check if the appropriate fields are provided
depending on the datalocation type and also check if the expected
access mechanism is used for the given datalocation.