Move Event authorization code to authorization module
As discussed on https://github.com/MIT-LCP/physionet-build/issues/2016,  we want to move all
authorization related logic to projects.authorization module to centralize authorization logic.

Note: here i created an events.py file inside authorization, because if we try to add this func
inside the projects.authorization.access.py, we will get circular import.
Also, this kind of organizes the authorization for event in a single file(instead of mixing up on access.py)

In future, it might make sense to move authorization as a separate app.