SubmissionInfo: add 'files' property.
We want to modularize project file storage, and in the future, use
different backends for different projects.  I'm not sure exactly how
this will look, but at a minimum, each part of the site that interacts
with project files will need to be aware of *which* project it is
dealing with.

Consequently, any function that needs a ProjectFiles instance must
obtain it from a project object (ActiveProject, PublishedProject, or
ArchivedProject); the ProjectFiles pseudo-constructor is deprecated
and should not be used.