Replace numerous uses of ProjectFiles with project.files.
We want to modularize project file storage, and in the future, use
different backends for different projects.  Functions that need a
ProjectFiles instance will need to obtain it from a project object
rather than calling the ProjectFiles pseudo-constructor.  Fix this for
existing cases where the usage is linked to an existing project.