Do not use the ActiveProject.version_order field.
The version_order field of the ActiveProject model (as well as
ArchivedProject) does not mean what it suggests.  At present, this
field is only set when the ActiveProject is created, and its only
significance is that it is zero for a new project (in which case
is_new_version is False) or non-zero for a new version (in which case
is_new_version is True).

(This is in contrast to the version_order field of *PublishedProject*,
which indicates the version sorting order of the published versions.
There is no real relationship between ActiveProject's version_order
and PublishedProject's version_order.)

The ActiveProject version_order field is therefore confusing and
redundant - everywhere it is currently used, is_new_version can and
should be used instead.

(Note that has_other_versions is not an ActiveProject field.)