validate_version: forbid leading zeroes in version numbers.
It should be possible to compare published project version numbers
using conventional version-number comparisons (as is done by
PublishedProject.set_version_order, for example.)  If leading zeroes
are allowed, it's impossible to tell whether "1.0.1" is greater or
less than "1.00.01".

(Version numbers are not decimal numbers.)

Disallow version numbers for new projects that don't follow this rule.