move_files_as_readonly: mark as an associated task.
When a project is published, certain tasks need to be performed
(setting read-only permissions and creating the list of SHA-256
hashes.)  These are handled by the background task
move_files_as_readonly which runs *after* the project is "published".
(This is not a great design and should be reconsidered.)

We should not be triggering other background tasks on the project
while this task is running (or if the task has been delayed for some
reason.)  In particular, we don't want to send files to cloud mirrors
or generate a zip archive until after the SHA256SUMS.txt file has been
generated.

Mark the move_files_as_readonly task as an "associated task" linked to
the PublishedProject, so that these actions will be disabled in the
admin console (see console.views.manage_published_project) while the
task is pending.