create_bucket: create the bucket in a single request.
When creating a GCS bucket, we need to set various bucket properties
in addition to the bucket name.  The
google.cloud.storage.Client.create_bucket method allows setting some
properties via keyword arguments.  However, the underlying HTTP API
allows setting all of the properties at once - which means we can
create the bucket and set its properties with a single request, rather
than making two requests and risking that the second one fails.
Note that the same pattern is used by physionet.gcs.create_bucket.

(It might be better if we also could set the IAM policy at the same
time - I'm not sure if this is possible or not - but on the other hand
it might be better *not* to make the bucket accessible until after the
project files have been uploaded.)