create_bucket: mark all buckets as requester-pays by default.
GCS buckets may be either "requester-pays" (the client accessing the
data must specify a "project" that will be billed for egress costs),
or "non-requester-pays" (any authorized client can access the data and
the bucket's owner will be billed for egress costs.)

This option can be switched on or off at any time, and it would be
nice to have a way to do so in the PhysioNet console.  For the time
being, we want to set all buckets as "requester-pays" by default.