files_panel: fix internal links and browser history.
In the files section of a preview or published project, clicking a
link to a subdirectory or parent directory should behave like a normal
link:

- The browser's window.location should be updated, so that you can
  bookmark it.

- The former location should be saved in the browser's history stack,
  so that you can navigate with Back and Forward buttons.

Doing these things requires calling history.pushState when we navigate
to a new location, and setting window.onpopstate to a handler function
that restores the former location.

Moreover, it is cleaner and safer to avoid using inline scripts.

When a directory link is clicked, we will asynchronously load the
directory contents; once they're loaded, call pushState to update the
page URL and then parse the new directory HTML.  (pushState must be
called first so that relative links are handled correctly.)

When the back button is clicked ("onpopstate" function is called), we
likewise load the previous HTML directory contents, but in this case
call replaceState instead of pushState (again, we need to update the
page URL before parsing the HTML.)  (You might think this replaceState
is unnecessary, but because the HTML is loaded asynchronously, there
could be multiple attempts in flight to change the page location, so
it's better to keep the DOM and page URL and history state all in sync
with each other.)

To avoid overly complicated transition logic, the existing
preview_files_panel and published_files_panel views are kept as-is.
Only when somebody loads the new project_preview page or
published_project page will they use the new script, and only when
using the new script (which adds a "v=2" query parameter) will
preview_files_panel and published_files_panel return the new
inline-script-free responses.

Eventually, preview_files_panel and published_files_panel can be
changed to return an error if the v=2 parameter is missing, and the
old files_panel.html can be removed.