console/urls.py: add parameters for TestURLs.
A few URLs in this app cannot currently be tested due to a lack of
demo data; in particular, testing the project submission pipeline
would require creating a demo project for each submission stage.

Many URLs in this app use a parameter called, uninformatively, "pk".
In most such cases, setting "pk" to 1 will work; but it'd be wiser to
use a descriptive name for the parameter.  It'd also be wiser not to
use integer IDs in URLs to begin with, and it would be wiser not to
assume that integer IDs are also primary keys.

Also note that user_access_logs_detail misleadingly names its
parameter "pid" (it's a user ID, not any kind of project ID.)