user.awsverification: new module.
This module contains functions for authenticating a person's AWS
identity (account number, username, and user ID) by means of signed
URLs.

Amazon S3 authenticates clients using a per-request "signature" that
incorporates the request path and headers together with a secret key
held by the client.  This means that the client can pre-compute this
signature and send it to someone else, allowing the recipient to
perform that request on that client's behalf, without revealing the
secret key itself.

We can arrange to create an S3 URL that can only be accessed by a
particular AWS identity, and then ask someone to pre-compute the
signature that they would use to access that resource (which they can
do using the AWS CLI or other S3-compatible tools and libraries.)

If we then submit that signature to S3 and it succeeds, we know that
the requester holds the secret key for that identity.  In fact, the
resource in question doesn't need to actually exist, as long as we can
tell the difference between an unauthorized request (HTTP 403) and an
authorized request for something that doesn't exist (HTTP 404).

AWS supports many types of identities.  This module only supports "IAM
user" identities (arn:aws:iam::*:user/*) and not any other types, both
as a matter of policy (data access permissions are granted to
individual people, not to groups, organizations, or computers) and
because IAM users have a stable, fixed-length user ID.