Add "IsDerivedFrom" to DOI metadata.
If a published project is derived from other published projects on the
site, include that information in the DOI metadata.

Formally, "IsDerivedFrom indicates B [the related resource] is a
source upon which A [the resource being registered] is based.
IsDerivedFrom should be used for a resource that is a derivative of
an original resource."

https://schema.datacite.org/meta/kernel-4.4/doc/DataCite-MetadataKernel_v4.4.pdf
(page 63)