Add model fields for storing AWS identity information.
We want to verify a person's AWS identity in order to permit them to
access restricted resources via direct cloud APIs; and possibly for
other purposes in the future.

An AWS account ID is not an identity.  An account may contain many
identities (known as "userids" or "unique IDs"), which might or might
not belong to the same person.  (Even if they do all belong to the
same person, it doesn't mean that person wants or should want to give
all of their identities the ability to access sensitive data.)

Here, we add fields to store the userid alongside the account ID and
ARN (which may be of interest in the future), and the date and time
that these credentials were verified.