create_s3_server_access_log_bucket: new function.
This function is used to create the bucket used for S3 server access
logs.  This function ordinarily will not be used (since the log bucket
only needs to be configured once), but is needed for testing.

Note that the log bucket requires a special access policy allowing the
AWS internal logging service to write logs on behalf of a specific
bucket owner (i.e. the account that owns the project buckets.)